requirements.txt
# 基础图片处理
Pillow>=10.0.1
opencv-python>=4.8.1.78
albumentations>=1.3.1

# 数值计算与数据处理
numpy>=1.24.3
pandas>=2.0.3
openpyxl>=3.1.2  # Excel导出依赖

# 深度学习框架（CPU版）
torch>=2.0.1+cpu
torchvision>=0.15.2+cpu
tensorflow>=2.13.0

# 网页开发
streamlit>=1.27.2

# 数据库连接
mysql-connector-python>=8.2.0

# 可视化
matplotlib>=3.7.2